requirements.txt
torch>=1.10
numpy
pandas
pyyaml
scikit-learn
